rename builder to message
