+++ src/message/encoder.rs
@@ -1,4 +1,4 @@
-use crate::builder::header::ContentTransferEncoding;
+use crate::message::header::ContentTransferEncoding;
 use base64;
 use bytes::{Buf, BufMut, Bytes, BytesMut, IntoBuf};
 use quoted_printable;
