+++ src/message/header/textual.rs
@@ -1,4 +1,4 @@
-use crate::builder::utf8_b;
+use crate::message::utf8_b;
 use hyperx::{
     header::{Formatter as HeaderFormatter, Header, RawLike},
     Error as HyperError, Result as HyperResult,
