+++ src/message/mod.rs
@@ -11,7 +11,7 @@
 The easiest way how we can create email message with simple string.
 
 ```rust
-use lettre::builder::Message;
+use lettre::message::Message;
 
 fn main() {
     let m: Message<&str> = Message::builder()
@@ -48,7 +48,7 @@ The more complex way is using MIME contents
 (see [format\_mime.rs](examples/format_mime.rs)).
 
 ```rust
-use lettre::builder::{header, Message, SinglePart};
+use lettre::message::{header, Message, SinglePart};
 
 fn main() {
     let m: Message<SinglePart<&str>> = Message::builder()
@@ -91,7 +91,7 @@ And more advanced way of building message by using multipart MIME contents
 (see [format\_multipart.rs](examples/format_multipart.rs)).
 
 ```rust
-use lettre::builder::{header, Message, MultiPart, SinglePart};
+use lettre::message::{header, Message, MultiPart, SinglePart};
 
 fn main() {
     let m: Message<MultiPart<&str>> = Message::builder()
